+++ forui/lib/src/theme/theme.dart
@@ -280,10 +280,7 @@ class FBasicTheme extends StatelessWidget {
       child: Directionality(
         textDirection: textDirection ?? Directionality.maybeOf(context) ?? .ltr,
         child: DefaultTextStyle(
-          style: data.typography.sm.copyWith(
-            fontFamily: data.typography.defaultFontFamily,
-            color: data.colors.foreground,
-          ),
+          style: data.typography.sm.copyWith(color: data.colors.foreground),
           child: child,
         ),
       ),