+++ forui/lib/src/widgets/item/item_mixin.dart
@@ -90,7 +90,7 @@ mixin FItemMixin on Widget {
     double? submenuCacheExtent,
     bool? submenuIntrinsicWidth,
     DragStartBehavior submenuDragStartBehavior = .start,
-    FPortalSpacing submenuSpacing = .zero,
+    FPortalSpacing submenuSpacing = const .spacing(2),
     FPortalOverflow submenuOverflow = .flip,
     Offset submenuOffset = .zero,
     String? submenuSemanticsLabel,