+++ forui/lib/src/widgets/tabs/tabs.dart
@@ -214,10 +214,7 @@ class _FTabsState extends State<FTabs> with SingleTickerProviderStateMixin {
     }
 
     final content = DefaultTextStyle(
-      style: theme.typography.md.copyWith(
-        fontFamily: theme.typography.defaultFontFamily,
-        color: theme.colors.foreground,
-      ),
+      style: theme.typography.md.copyWith(color: theme.colors.foreground),
       child: useTabBarView
           ? TabBarView(
               controller: _controller._controller,